requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
gitpython>=3.1.40
pydantic>=2.5.0
requests>=2.31.0
playwright>=1.40.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
jinja2>=3.1.0
python-multipart>=0.0.6
